GameDTO, GameRepository, GameService, GameController

+++ src/main/java/br/com/teixeira/dslist/entities/Game.java
@@ -13,7 +13,7 @@ public class Game {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String title;
 
     @Column(name = "game_year")
@@ -31,7 +31,7 @@ public class Game {
 
     }
 
-    public Game(Integer id, String title, Integer year, String genre, String platforms, Double score,String imgUrl,
+    public Game(Long id, String title, Integer year, String genre, String platforms, Double score,String imgUrl,
             String shortDescription, String longDescription) {
                 
         this.id = id;
@@ -45,11 +45,11 @@ public class Game {
         this.longDescription = longDescription;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -141,6 +141,8 @@ public class Game {
             return false;
         return true;
     }
+
+    
 
 }
